Main background component
