+++ frontend/app/pages/SettingsPage.tsx
@@ -1,12 +1,12 @@
 import { useState } from "react";
-import { View, Text, ImageBackground, TouchableOpacity } from "react-native";
-import background from "../../assets/background.png";
+import { View, Text, TouchableOpacity } from "react-native";
 
 import AboutSection from "./settings/About";
 import CreditsSection from "./settings/Credits";
 import HelpSection from "./settings/Help";
 import SettingsMainSection from "./settings/Main";
 import { ClaimRewardSection } from "./settings/ClaimReward";
+import MainBackground from "../components/MainBackground";
 
 const tabs = {
   Main: SettingsMainSection,
@@ -22,23 +22,22 @@ export const SettingsPage: React.FC = () => {
   const ActiveComponent = tabs[activeTab];
 
   return (
-    <ImageBackground
-      className="flex-1 flex flex-col gap-2 px-8"
-      source={background}
-      resizeMode="cover"
-    >
-      <ActiveComponent setSettingTab={setActiveTab} />
-      {activeTab !== "Main" && (
-        <TouchableOpacity
-          onPress={() => {
-            setActiveTab("Main");
-          }}
-          className="bg-[#f0a030] p-4 rounded-xl border-2 border-[#ffffff80] flex flex-row justify-center items-center"
-        >
-          <Text className="text-4xl">Back to Settings ⚙️</Text>
-        </TouchableOpacity>
-      )}
-    </ImageBackground>
+    <View className="flex-1 relative">
+      <MainBackground />
+      <View className="flex flex-col gap-2 px-8">
+        <ActiveComponent setSettingTab={setActiveTab} />
+        {activeTab !== "Main" && (
+          <TouchableOpacity
+            onPress={() => {
+              setActiveTab("Main");
+            }}
+            className="bg-[#f0a030] p-4 rounded-xl border-2 border-[#ffffff80] flex flex-row justify-center items-center"
+          >
+            <Text className="text-4xl">Back to Settings ⚙️</Text>
+          </TouchableOpacity>
+        )}
+      </View>
+    </View>
   );
 };
